Fix contention issue on extraction folder

+++ Quality/Function-Extraction.Tests.ps1
@@ -3,14 +3,22 @@ param(
     [string[]]$SourcePath,
 
     [parameter(Mandatory = $true)]
-    [string[]]$FunctionExtractPath
+    [string]$FunctionExtractPath
 )
 
 Describe "Function Extraction Tests" -Tag 'Setup' {
 
-    Get-ChildItem -Path $FunctionExtractPath -Recurse | Remove-Item -Force -Recurse
-    Remove-Item $FunctionExtractPath -Force -ErrorAction SilentlyContinue
-    New-Item -Path $FunctionExtractPath -ItemType 'Directory'
+    if ( -not (Test-Path -Path $FunctionExtractPath)) {
+        # New-Item -Path $extractPath.Parent -ItemType 'Directory'
+        New-Item -Path $FunctionExtractPath -ItemType 'Directory'
+    }
+    else {
+        $extractPath = Get-Item -Path $FunctionExtractPath
+        Get-ChildItem -Path $FunctionExtractPath -Recurse -ErrorAction SilentlyContinue | Remove-Item -Force -Recurse
+        Remove-Item $FunctionExtractPath -Force -ErrorAction SilentlyContinue
+        # New-Item -Path $extractPath.Parent -ItemType 'Directory'
+        New-Item -Path $FunctionExtractPath -ItemType 'Directory'
+    }
 
     foreach ($scriptPath in $SourcePath) {
 
@@ -24,7 +32,7 @@ Describe "Function Extraction Tests" -Tag 'Setup' {
 
                     {
 
-                        Export-FunctionsFromModule -Path $moduleFile
+                        Export-FunctionsFromModule -Path $moduleFile -FunctionExtractPath $FunctionExtractPath
 
                     } | Should -Not -Throw
